+++ parapet-cats-effect/src/test/scala/io/parapet/tests/intg/cats/CatsEffectSpecs.scala
@@ -2,21 +2,20 @@ package io.parapet.tests.intg.cats
 
 import cats.effect.IO
 import io.parapet.cats.CatsEffectParApp
-import io.parapet.core.Parapet
 import io.parapet.core.processes.DeadLetterProcess
 import io.parapet.journal.EventCodecRegistry
 import io.parapet.tests.intg.IntegrationSpec
-import io.parapet.{ParApp, core}
+import io.parapet.{ParApp, ParConfig, core}
 
 trait BasicCatsEffectSpec extends IntegrationSpec[IO] with CatsEffectParApp:
   override def createApp(
       processes0: IO[Seq[core.Process[IO, ?, ?]]],
       deadLetter0: Option[IO[DeadLetterProcess[IO]]],
-      config0: Parapet.ParConfig,
+      config0: ParConfig,
       eventCodecs0: EventCodecRegistry
   ): ParApp[IO] =
     new CatsEffectParApp:
-      override val config: Parapet.ParConfig = config0
+      override val config: ParConfig = config0
 
       override def processes(args: Array[String]): IO[Seq[core.Process[IO, ?, ?]]] =
         processes0
