+++ parapet-core/src/test/scala/io/parapet/core/TestUtils.scala
@@ -1,7 +1,7 @@
 package io.parapet.core
 
 import io.parapet.core.Dsl.*
-import io.parapet.core.Parapet.ParConfig
+import io.parapet.ParConfig
 import io.parapet.runtime.Scheduler.{Deliver, SubmissionResult, Task}
 import io.parapet.core.processes.Noop
 import io.parapet.effect.{Effect, EffectFiber, Monad}
