+++ parapet-testkit/src/main/scala/io/parapet/tests/intg/BlockingSpec.scala
@@ -1,7 +1,7 @@
 package io.parapet.tests.intg
 
 import io.parapet.core.Events.{DeadLetter, Kill, Start, Stop}
-import io.parapet.core.Parapet.ParConfig
+import io.parapet.ParConfig
 import io.parapet.core.Process
 import io.parapet.runtime.Scheduler.SchedulerConfig
 import io.parapet.core.exceptions.EventHandlingException
