+++ parapet-testkit/src/main/scala/io/parapet/tests/intg/ChannelSpec.scala
@@ -2,7 +2,7 @@ package io.parapet.tests.intg
 
 import io.parapet.core.Dsl.DslF
 import io.parapet.core.Events.{Start, Stop}
-import io.parapet.core.Parapet.ParConfig
+import io.parapet.ParConfig
 import io.parapet.runtime.Scheduler.SchedulerConfig
 import io.parapet.core.exceptions.EventHandlingException
 import io.parapet.core.{Channel, Process}
